Update template.sql
release_date made optional because it must be possible to create a project without knowing the it beforehand. I'm also taking a wild guess regarding project.contract that it might not be known the moment the project is created, made optional as well. Added trigger for project.revision to increment it every time revision_date is updated. Now counting from 1 instead of 0 everywhere - feels more logical regarding the type of data we're dealing with. Added trigger constraint to project to limit table to one row.